\name{levelplotSNP}
\alias{levelplotSNP}
\title{Gene correlation matrix heatmap}
\usage{
levelplotSNP(Sigma)
}
\arguments{
\item{Sigma}{a correlation matrix}
}
\description{
\code{levelplotSNP} is used to plot a correlation matrix as an heatmap. It is based on the \code{levelplot}
function from the \code{lattice} package.
}
\references{
Sarkar, Deepayan (2008) Lattice: Multivariate Data Visualization with R. Springer, New York.
ISBN 978-0-387-75968-5
}
