% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationSNPSet.R
\name{PopulationSNPSet}
\alias{PopulationSNPSet}
\title{Simulation of a population of genotypic profiles}
\usage{
PopulationSNPSet(n, Sigma = NULL, maf = NULL, marginal = NULL,
  X = NULL)
}
\arguments{
\item{n}{the desired number of individuals}

\item{Sigma}{a correlation matrix}

\item{maf}{a vector of minor allele frequencies; if this is given, Hardy-Weinberg equilibrium is assumed}

\item{marginal}{a list of marginal distributions, where each element of the list is a vector of length 1 or 2
of cumulative probabilities. See the \code{ordsample} function from the \code{GenOrd} package for more
information}

\item{X}{an observed matrix of genotypic profiles; if this is given, the sample is simulated according to
the dependence structure and marginal distributions observed in \code{X}.}
}
\value{
A matrix \code{X} of size N x p of genotypic profiles (displayed as rows).
}
\description{
\code{PopulationSNPSet} is used to generate a population matrix, where each row is a genotypic profile
corresponding to an individual. It can be used to generate a matrix of genotypes according to an observed
dependence structure. This function is based on the \code{GenOrd} package.
}
\references{
Alessandro Barbiero and Pier Alda Ferrari (2015). GenOrd: Simulation of Discrete Random Variables
with Given Correlation Matrix and Marginal Distributions. R package version 1.4.0.
https://CRAN.R-project.org/package=GenOrd
}
